+++ KSP2Runtime/KSPVessel/KSPVesselModule.ModuleTransmitter.cs
@@ -7,8 +7,7 @@ namespace KontrolSystem.KSP.Runtime.KSPVessel;
 public partial class KSPVesselModule {
     [KSClass("ModuleTransmitter")]
     public class ModuleTransmitterAdapter : BaseTransmitterAdapter<PartAdapter, PartComponent> {
-        public ModuleTransmitterAdapter(PartAdapter part, Data_Transmitter dataTransmitter) : base(part, dataTransmitter)
-        {
+        public ModuleTransmitterAdapter(PartAdapter part, Data_Transmitter dataTransmitter) : base(part, dataTransmitter) {
         }
     }
 }